+++ test_autofit/mapper/with_paths/test_model.py
@@ -1,28 +1,6 @@
 import pytest
 
-import autofit as af
 from autofit.mapper.prior_model.abstract import paths_to_tree
-
-
-@pytest.fixture(
-    name="gaussian_1"
-)
-def make_gaussian_1():
-    return af.Model(
-        af.Gaussian
-    )
-
-
-@pytest.fixture(
-    name="model"
-)
-def make_model(gaussian_1):
-    return af.Collection(
-        gaussian_1=gaussian_1,
-        gaussian_2=af.Model(
-            af.Gaussian
-        ),
-    )
 
 
 class TestWithoutAttributes:
