+++ test_autofit/mapper/with_paths/test_samples.py
@@ -1,0 +1,5 @@
+import autofit as af
+
+
+def test_sample():
+    pass
